+++ aws/cloudtrail/examples/complete/main.tf
@@ -1,0 +1,7 @@
+module "cloudtrail" {
+  source = "../../"
+
+  name                          = "example-cloudtrail"
+  enable_logging                = true
+  include_global_service_events = true
+}