+++ aws/kms/iam.tf
@@ -1,4 +1,4 @@
-data "aws_iam_policy_document" "encrypt_decrypt" {
+data "aws_iam_policy_document" "this" {
   statement {
     effect = "Allow"
 
@@ -17,5 +17,24 @@ data "aws_iam_policy_document" "encrypt_decrypt" {
 
 resource "aws_kms_key_policy" "this" {
   key_id = aws_kms_key.this.id
+  policy = data.aws_iam_policy_document.encrypt_decrypt.json
+}
+
+data "aws_iam_policy_document" "encrypt_decrypt" {
+  statement {
+    effect = "Allow"
+    actions = [
+      "kms:Encrypt",
+      "kms:Decrypt"
+    ]
+    resources = [aws_kms_key.this.arn]
+  }
+}
+
+
+resource "aws_iam_policy" "this" {
+  for_each = toset(var.alias_name)
+  name     = "KMSEncryptDecrypt-${each.value}"
+
   policy = data.aws_iam_policy_document.encrypt_decrypt.json
 }