+++ aws/postgres/variables.tf
@@ -4,6 +4,12 @@ variable "db_name" {
   type        = string
 }
 
+variable "db_identifier" {
+  description = "The identifier name of database instance. If null, the db_name will be used instead."
+  type        = string
+  default     = null
+}
+
 variable "db_username" {
   description = "The master username for the database."
   type        = string
@@ -65,6 +71,13 @@ variable "multi_az" {
   description = "Indicates whether the database instance should be deployed across multiple availability zones."
   type        = bool
   default     = false
+}
+
+# Custom parameter group
+variable "custom_parameter_group_name" {
+  description = "Custom parameter group name, used when `var.overwrite_parameter_group_name` is `true` and `var.supported_engine_version` size is 1."
+  type        = string
+  default     = null
 }
 
 # Security
@@ -147,4 +160,11 @@ variable "supported_engine_version" {
   description = "A list of supported engine versions for the Parameter Groups, supporting Blue-Green deployment."
   type        = list(number)
   default     = [14, 15, 16]
+}
+
+# Security groups
+variable "vpc_security_group_ids" {
+  description = "The list of existing vpc security group ids to associate with database instance."
+  type        = list(string)
+  default     = null
 }