+++ aws/route53/examples/complete/outputs.tf
@@ -9,3 +9,7 @@ output "r53_main_zone_name" {
 output "first_record_fqdn" {
   value = module.route53.first_record_fqdn
 }
+
+output "r53_main_name_servers" {
+  value = module.route53.r53_main_name_servers
+}