+++ aws/route53/outputs.tf
@@ -12,3 +12,8 @@ output "first_record_fqdn" {
   description = "Fully qualified domain name (FQDN) of the first Route 53 record, if available"
   value       = length(aws_route53_record.main) > 0 ? values(aws_route53_record.main)[0].fqdn : null
 }
+
+output "r53_main_name_servers" {
+  description = "A list of name servers in associated (or default) delegation set."
+  value       = try(aws_route53_zone.dns_zone[0].name_servers, data.aws_route53_zone.dns_zone[0].name_servers)
+}