+++ aws/vpc/main.tf
@@ -13,8 +13,7 @@ resource "aws_vpc" "this" {
   enable_dns_hostnames = true
 
   tags = {
-    Name        = "${var.name}-vpc"
-    Environment = var.environment
+    Name = "${var.name}-vpc"
   }
 }
 
@@ -26,8 +25,7 @@ resource "aws_internet_gateway" "main" {
   vpc_id = aws_vpc.this.id
 
   tags = {
-    Name        = "${var.name}-igw"
-    Environment = var.environment
+    Name = "${var.name}-igw"
   }
 }
 
@@ -42,8 +40,7 @@ resource "aws_nat_gateway" "main" {
   depends_on    = [aws_internet_gateway.main]
 
   tags = {
-    Name        = "${var.name}-nat-${format("%03d", count.index + 1)}"
-    Environment = var.environment
+    Name = "${var.name}-nat-${format("%03d", count.index + 1)}"
   }
 }
 
@@ -55,8 +52,7 @@ resource "aws_eip" "nat" {
   count = var.single_nat ? 1 : length(var.availability_zone_postfixes)
 
   tags = {
-    Name        = "${var.name}-eip-${format("%03d", count.index + 1)}"
-    Environment = var.environment
+    Name = "${var.name}-eip-${format("%03d", count.index + 1)}"
   }
 }
 
@@ -65,14 +61,14 @@ aws_subnet creates private subnets within the VPC for internal resources without
 https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/subnet
 */
 resource "aws_subnet" "private" {
+  count = length(var.availability_zone_postfixes)
+
   vpc_id            = aws_vpc.this.id
-  count             = length(var.availability_zone_postfixes)
-  cidr_block        = cidrsubnet(local.private_cidr_blocks, 8, count.index)
+  cidr_block        = var.create_custom_subnets ? element(var.custom_private_subnets, count.index) : cidrsubnet(local.private_cidr_blocks, 8, count.index)
   availability_zone = "${var.region}${element(var.availability_zone_postfixes, count.index)}"
 
   tags = {
-    Name        = "${var.name}-private-subnet-${format("%03d", count.index + 1)}"
-    Environment = var.environment
+    Name = "${var.name}-private-subnet-${format("%03d", count.index + 1)}"
   }
 }
 
@@ -81,15 +77,15 @@ aws_subnet creates public subnets within the VPC with direct internet access.
 https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/subnet
 */
 resource "aws_subnet" "public" {
+  count = length(var.availability_zone_postfixes)
+
   vpc_id                  = aws_vpc.this.id
-  count                   = length(var.availability_zone_postfixes)
-  cidr_block              = cidrsubnet(local.public_cidr_blocks, 8, count.index)
+  cidr_block              = var.create_custom_subnets ? element(var.custom_public_subnets, count.index) : cidrsubnet(local.public_cidr_blocks, 8, count.index)
   availability_zone       = "${var.region}${element(var.availability_zone_postfixes, count.index)}"
   map_public_ip_on_launch = true
 
   tags = {
-    Name        = "${var.name}-public-subnet-${format("%03d", count.index + 1)}"
-    Environment = var.environment
+    Name = "${var.name}-public-subnet-${format("%03d", count.index + 1)}"
   }
 }
 
@@ -101,8 +97,7 @@ resource "aws_route_table" "public" {
   vpc_id = aws_vpc.this.id
 
   tags = {
-    Name        = "${var.name}-routing-table-public"
-    Environment = var.environment
+    Name = "${var.name}-routing-table-public"
   }
 }
 
@@ -125,8 +120,7 @@ resource "aws_route_table" "private" {
   vpc_id = aws_vpc.this.id
 
   tags = {
-    Name        = "${var.name}-routing-table-private-${format("%03d", count.index + 1)}"
-    Environment = var.environment
+    Name = "${var.name}-routing-table-private-${format("%03d", count.index + 1)}"
   }
 }