+++ aws/vpc/variables.tf
@@ -8,11 +8,6 @@ variable "region" {
   type        = string
 }
 
-variable "environment" {
-  description = "The environment of this VPC"
-  type        = string
-}
-
 variable "cidr_block" {
   description = "The CIDR block for the VPC."
   type        = string
@@ -27,4 +22,22 @@ variable "single_nat" {
   description = "Whether to create a single NAT gateway or one per AZ"
   type        = bool
   default     = false
+}
+
+variable "create_custom_subnets" {
+  description = "Whether to create custom subnets."
+  type        = bool
+  default     = false
+}
+
+variable "custom_public_subnets" {
+  description = "List of custom public subnets."
+  type        = list(string)
+  default     = []
+}
+
+variable "custom_private_subnets" {
+  description = "List of custom private subnets."
+  type        = list(string)
+  default     = []
 }