+++ aws/route53/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.1.14]() (2024-12-18)
+### Features
+* Add `r53_main_name_servers` output for route53
+
+## [0.1.12]() (2024-12-17)
+### Features
+* Add `r53_main_name_servers` output for route53
+
 ## [0.1.4]() (2024-12-05)
 ### Features
 * Update terraform version constraint from `~> 1.9.8` to `>= 1.9.8`