+++ aws/s3/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.1.12]() (2024-12-17)
+### Features
+* Add attribute `force_destroy` for s3 bucket 
+* Add `read_write_policy`
+
 ## [0.1.4]() (2024-12-05)
 ### Features
 * Update terraform version constraint from `~> 1.9.8` to `>= 1.9.8`